move logic to the ContextMenuControllableTextField
