+++ wire-ios/Wire-iOS/Sources/UserInterface/Conversation/Create/WireTextField.swift
@@ -20,7 +20,7 @@ import UIKit
 import WireCommonComponents
 import WireDesign
 
-class WireTextField: UITextField {
+class WireTextField: ContextMenuControllableTextField {
 
     // MARK: - Properties
 
@@ -28,7 +28,6 @@ class WireTextField: UITextField {
         case valid(String)
         case error(SimpleTextFieldValidator.ValidationError)
     }
-    private let isContextMenuAllowed: Bool
     private let borderWidth: CGFloat = 1
     private let cornerRadius: CGFloat = 12
 
@@ -51,8 +50,7 @@ class WireTextField: UITextField {
     // MARK: - Initialization
 
     init(frame: CGRect, isContextMenuAllowed: Bool) {
-        self.isContextMenuAllowed = isContextMenuAllowed
-        super.init(frame: frame)
+        super.init(isContextMenuAllowed: isContextMenuAllowed)
         setup()
     }
 
@@ -185,31 +183,4 @@ extension WireTextField: SimpleTextFieldValidatorDelegate {
     func textFieldDidBeginEditing() {
         wireTextFieldDelegate?.textFieldDidBeginEditing(self)
     }
-}
-
-extension WireTextField {
-
-    override func canPerformAction(
-        _ action: Selector,
-        withSender sender: Any?
-    ) -> Bool {
-        if !isContextMenuAllowed {
-            let validActions = [
-                #selector((any UIResponderStandardEditActions).select(_:)),
-                #selector((any UIResponderStandardEditActions).selectAll(_:))
-            ]
-            return text!.isEmpty ? false : validActions.contains(action)
-        } else {
-            return super.canPerformAction(action, withSender: sender)
-        }
-    }
-
-    override func buildMenu(with builder: any UIMenuBuilder) {
-        if !isContextMenuAllowed {
-            if #available(iOS 17.0, *) {
-                builder.remove(menu: .autoFill)
-            }
-        }
-    }
-
 }
